Load reserve once in GrantW::commit_inner
This is testing my own understanding of how this is working.  I was looking at porting a variation of this to C++, when I noticed that the return value of reserve in GrantW::commit_inner is updated in fetch_sub, but the return value of the fetch_sub is not used, forcing another load.  This seemed like it could cause problems if the value was updated between the fetch_sub and the load.